U-GAT-IT/requirements.txt
numpy~=1.18.1
torchvision~=0.5.0
pillow~=8.1.1
tqdm~=4.47.0
opencv-python~=4.2.0.34
easydict~=1.9